+++ proxy/types/election.go
@@ -4,22 +4,6 @@ import (
 	etypes "github.com/dedis/d-voting/contracts/evoting/types"
 )
 
-// TransactionStatus is the status of a transaction
-type TransactionStatus byte
-
-const (
-	// UnknownTransactionStatus is the basic status of a transaction
-	UnknownTransactionStatus TransactionStatus = 0
-	// IncludedTransaction is the status of a transaction that has been included
-	IncludedTransaction TransactionStatus = 1
-	// RejectedTransaction is the status of a transaction will never be included
-	RejectedTransaction TransactionStatus = 2
-)
-
-
-
-
-
 // CreateFormRequest defines the HTTP request for creating a form
 type CreateFormRequest struct {
 	AdminID       string
@@ -29,7 +13,7 @@ type CreateFormRequest struct {
 // CreateFormResponse defines the HTTP response when creating a form
 type CreateFormResponse struct {
 	FormID string // hex-encoded
-	Token string
+	Token  string
 }
 
 // CastVoteRequest defines the HTTP request for casting a vote
@@ -38,7 +22,6 @@ type CastVoteRequest struct {
 	// Marshalled representation of Ciphervote. It contains []{K:,C:}
 	Ballot CiphervoteJSON
 }
-
 
 // CiphervoteJSON is the JSON representation of a ciphervote
 type CiphervoteJSON []EGPairJSON
@@ -49,29 +32,6 @@ type EGPairJSON struct {
 	C []byte
 }
 
-// TransactionInfo defines the information of a transaction 
-type TransactionInfo struct {
-	Status TransactionStatus // 0 if not yet included, 1 if included, 2 if rejected
-	TransactionID []byte
-	LastBlockIdx uint64  // last block of the chain when the transaction was added to the pool
-	Time 	   int64 // time when the transaction was added to the pool
-	Hash    []byte // signature of the transaction
-	Signature []byte // signature of the transaction
-}
-
-// TransactionInfoToSend defines the HTTP response when sending 
-// transaction infos to the client so that he can use the status
-// of the transaction to know if it has been included or not
-// and if it has not been included, he can just use the token
-// and ask again later
-type TransactionInfoToSend struct {
-	Status TransactionStatus // 0 if not yet included, 1 if included, 2 if rejected
-	Token string
-}
-
-
-
-
 // UpdateFormRequest defines the HTTP request for updating a form
 type UpdateFormRequest struct {
 	Action string
@@ -80,7 +40,7 @@ type UpdateFormRequest struct {
 // GetFormResponse defines the HTTP response when getting the form info
 type GetFormResponse struct {
 	// FormID is hex-encoded
-	FormID      string
+	FormID          string
 	Configuration   etypes.Configuration
 	Status          uint16
 	Pubkey          string
@@ -94,9 +54,9 @@ type GetFormResponse struct {
 // LightForm represents a light version of the form
 type LightForm struct {
 	FormID string
-	Title      string
-	Status     uint16
-	Pubkey     string
+	Title  string
+	Status uint16
+	Pubkey string
 }
 
 // GetFormsResponse defines the HTTP response when getting all forms
